LayoutTests/imported/w3c/web-platform-tests/cors/preflight-failure-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%266&code=301&preflight=301&7. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%264&code=300&preflight=300&5. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%268&code=302&preflight=302&9. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2610&code=303&preflight=303&11. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2612&code=304&preflight=304&13. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2614&code=305&preflight=305&15. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2618&code=307&preflight=307&19. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2620&code=308&preflight=308&21. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2616&code=306&preflight=306&17. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2624&code=401&preflight=401&25 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2622&code=400&preflight=400&23. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2626&code=402&preflight=402&27. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2628&code=403&preflight=403&29. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2632&code=405&preflight=405&33. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2630&code=404&preflight=404&31. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2634&code=406&preflight=406&35. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2636&code=407&preflight=407&37. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2638&code=408&preflight=408&39. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2640&code=409&preflight=409&41. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2642&code=410&preflight=410&43. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2644&code=411&preflight=411&45. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2646&code=412&preflight=412&47. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2648&code=413&preflight=413&49. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2650&code=414&preflight=414&51. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2652&code=415&preflight=415&53. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2654&code=416&preflight=416&55. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2656&code=417&preflight=417&57. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2658&code=500&preflight=500&59. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2660&code=501&preflight=501&61. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2662&code=502&preflight=502&63. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2664&code=503&preflight=503&65. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2666&code=504&preflight=504&67. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2668&code=505&preflight=505&69. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2670&code=680&preflight=680&71. Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?headers=x-test&location=http%3A%2F%2F127.0.0.1%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3Fheaders%3Dx-test%2672&code=790&preflight=790&73. Preflight response is not successful
Preflight responds with non-2XX status code


PASS Should succeed if preflight has status 200 
PASS Should succeed if preflight has status 299 
PASS Should throw error if preflight has status 300 
PASS Should throw error if preflight has status 301 
PASS Should throw error if preflight has status 302 
PASS Should throw error if preflight has status 303 
PASS Should throw error if preflight has status 304 
PASS Should throw error if preflight has status 305 
PASS Should throw error if preflight has status 306 
PASS Should throw error if preflight has status 307 
PASS Should throw error if preflight has status 308 
PASS Should throw error if preflight has status 400 
PASS Should throw error if preflight has status 401 
PASS Should throw error if preflight has status 402 
PASS Should throw error if preflight has status 403 
PASS Should throw error if preflight has status 404 
PASS Should throw error if preflight has status 405 
PASS Should throw error if preflight has status 406 
PASS Should throw error if preflight has status 407 
PASS Should throw error if preflight has status 408 
PASS Should throw error if preflight has status 409 
PASS Should throw error if preflight has status 410 
PASS Should throw error if preflight has status 411 
PASS Should throw error if preflight has status 412 
PASS Should throw error if preflight has status 413 
PASS Should throw error if preflight has status 414 
PASS Should throw error if preflight has status 415 
PASS Should throw error if preflight has status 416 
PASS Should throw error if preflight has status 417 
PASS Should throw error if preflight has status 500 
PASS Should throw error if preflight has status 501 
PASS Should throw error if preflight has status 502 
PASS Should throw error if preflight has status 503 
PASS Should throw error if preflight has status 504 
PASS Should throw error if preflight has status 505 
PASS Should throw error if preflight has status 680 
PASS Should throw error if preflight has status 790 
